% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_optimal_values.R
\name{get_optimal_values}
\alias{get_optimal_values}
\title{Hauspreisrechner}
\usage{
get_optimal_values(
  V,
  E,
  r,
  u,
  alpha,
  min_eigenmittel,
  amortisation_jahre,
  amortisation_grenze,
  z0
)
}
\arguments{
\item{V}{Vermögen}

\item{E}{Einkommen}

\item{r}{Zinssatz der Hypothek}

\item{u}{Prozentsatz Unterhaltskosten}

\item{alpha}{maximaler Einkommensanteil zur Bedienung der Hypothek}

\item{min_eigenmittel}{Minimale Eigenmittel in Prozent}

\item{amortisation_jahre}{Anzal Jahre bis zur minimalen Amortisation}

\item{amortisation_grenze}{maximale Belehnung nach X Jahren}

\item{z0}{Startwert für Vermögensanteil z, über welchen iterriert wird}
}
\value{
Eine Liste mit dem optimalen y und z pro Maximierungsproblem. Dazu
kommen für beide Maximierungsprobleme die Werte, wenn die y und z in die
jeweiligen Nebenbedingungen eingesetzt werden.
}
\description{
Basierend auf dem Vermögen (V) und dem Einkommen (E) wird mit dieser Funktion
der maximal leistbare Hauspreis bestimmt. In einer Vielzahl von Fällen lässt
sich der maximale Hauspreis mit einfach Heuristiken herleiten. Wo dies nicht
der Fall ist, wird das optimale y und z mit Hilfer der Funktion
\code{solve_max_problem} berechnet.
}
