% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_optimal_values.R
\name{get_optimal_y}
\alias{get_optimal_y}
\title{Get optimal y}
\usage{
get_optimal_y(
  V,
  E,
  r,
  u,
  alpha,
  min_eigenmittel,
  amortisation_jahre,
  amortisation_grenze,
  z0
)
}
\arguments{
\item{V}{Vermögen}

\item{E}{Einkommen}

\item{r}{Zinssatz der Hypothek}

\item{u}{Prozentsatz Unterhaltskosten}

\item{alpha}{maximaler Einkommensanteil zur Bedienung der Hypothek}

\item{min_eigenmittel}{Minimale Eigenmittel in Prozent}

\item{amortisation_jahre}{Anzal Jahre bis zur minimalen Amortisation}

\item{amortisation_grenze}{maximale Belehnung nach X Jahren}

\item{z0}{Startwert für Vermögensanteil z, über welchen iterriert wird}
}
\value{
Der maximale Kaufpreis eines Hauses
}
\description{
Get optimal y
}
