% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{solve_max_problem}
\alias{solve_max_problem}
\title{maximaler Hauspreis solver 2}
\usage{
solve_max_problem(
  V,
  E,
  r = 0.05,
  u = 0.007,
  alpha = 0.35,
  min_eigenmittel = 0.2,
  amortisation_jahre = 15,
  amortisation_grenze = 2/3,
  z0 = 0.5
)
}
\arguments{
\item{V}{Vermögen}

\item{E}{Einkommen}

\item{r}{Zinssatz der Hypothek}

\item{u}{Prozentsatz Unterhaltskosten}

\item{alpha}{maximaler Einkommensanteil zur Bedienung der Hypothek}

\item{min_eigenmittel}{Minimale Eigenmittel in Prozent}

\item{amortisation_jahre}{Anzal Jahre bis zur minimalen Amortisation}

\item{amortisation_grenze}{maximale Belehnung nach X Jahren}

\item{z0}{Startwert für Vermögensanteil z, über welchen iterriert wird}
}
\value{
Eine Liste mit dem optimalen y und z pro Maximierungsproblem. Dazu
kommen für beide Maximierungsprobleme die Werte, wenn die y und z in die
jeweiligen Nebenbedingungen eingesetzt werden.
}
\description{
Ziel: Finde ein optimales z, welches den Kaufpreis y maximiert. Dabei stellt
z der Anteil des eigenen Vermögens dar, welcher von der Bank vervielfacht
wird (--> 0 <= z <= 1).
}
